backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2

# Génération de documents PDF
docxtpl==0.19.0
python-docx==1.1.2
reportlab==4.2.2
weasyprint==62.3
Pillow==10.4.0
django-filter==24.3

# Dépendances supplémentaires pour la génération PDF
jinja2==3.1.4
lxml==5.1.0
cffi==1.16.0
pycparser==2.21
html5lib==1.1
tinycss2>=1.3.0
cssselect2==0.7.0
pyphen==0.14.0
cairocffi==1.6.1
cairosvg==2.7.1

#pour le chatbot
spacy==3.7.4
requests==2.32.3
fr-core-news-md==3.7.0

# Text-to-SQL avec LangChain pour la generation automatique des requetes
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
openai>=1.40.0

# Django Channels pour les WebSockets et notifications temps réel
channels==4.1.0
channels-redis==4.2.0
redis==5.0.1
daphne==4.2.1
websockets==15.0.1

# Gestion des fichiers et uploads
Pillow==10.4.0